rename: 모듈 파일 경로 변경
