+++ src/presentation/controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { NotificationController } from './notification.controller';
+import { NotificationController } from './controller';
 
 describe('NotificationController', () => {
   let controller: NotificationController;
